[重構] 後端 getProducts
