+++ lib/product_tools.js
@@ -27,7 +27,26 @@ module.exports = {
     return queryString
   },
 
-  setWhere(req, where) {
+  getOrder(req) {
+    // 排序條件，預設為 ['releaseDate', 'DESC']
+    const sort = req.query.sort || 'releaseDate'
+    const orderBy = req.query.order || 'DESC'
+    const order = [[sort, orderBy]]
+    const selectedSort = `${sort},${orderBy}`
+
+    return [order, selectedSort]
+  },
+
+  setWhere(req, now) {
+    // 設定 where default
+    const where = {
+      status: 1,
+      releaseDate: { [Op.lte]: now },
+      [Op.and]: [  // 搭配 searchQuery 組雙 OR
+        { [Op.or]: { deadline: { [Op.gte]: now }, saleDate: { [Op.lte]: now } } }
+      ]
+    }
+
     // 製作 category where
     let categoryQuery = +req.query.category || 0
     if (categoryQuery) { where.CategoryId = categoryQuery }
@@ -53,13 +72,11 @@ module.exports = {
 
     if (tagQuery === 'soon') {
       // 從 Date.now 往後取 7 天快截止的商品
-      const now = new Date()
       const end = moment().add(7, 'days').endOf('day')
-
       where.deadline = { [Op.between]: [now, end] }
     }
 
-    return { categoryQuery, searchQuery, tagQuery }
+    return { where, categoryQuery, searchQuery, tagQuery }
   },
 
   getPagination(products, PAGE_LIMIT, page) {
